Add separate config for all variables
* Add a script for cloning repos if needed
* Fix ci_url
* Fix dowwnload url which is sent to tg
* Use "" everywhere

+++ build.sh
@@ -1,13 +1,11 @@
 #!/bin/bash
 
-# Email for git
-GITHUB_USER=Akianonymus
-ci_repo=$(cat /drone/src/.git/config | grep url | sed 's|url = https://github.com/||' | sed 's|.git||')
-git config --global user.email "anonymus.aki@gmail.com"
-git config --global user.name "$GITHUB_USER"
+source /drone/src/config.sh
 
-export KBUILD_BUILD_USER="Aki"
-export KBUILD_BUILD_HOST="A_DEAD_PLANET"
+# Email for git
+ci_repo=$(cat /drone/src/.git/config | grep url | sed 's|url = https://github.com/||' | sed 's|.git||')
+git config --global user.email "$GITHUB_EMAIL"
+git config --global user.name "$GITHUB_USER"
 
 TELEGRAM_TOKEN=$(cat /tmp/tg_token)
 TELEGRAM_CHAT=$(cat /tmp/tg_chat)
@@ -16,11 +14,6 @@ GITHUB_TOKEN=$(cat /tmp/gh_token)
 export TELEGRAM_TOKEN
 export TELEGRAM_CHAT
 export GITHUB_TOKEN
-
-function set_device() {
-    device="harpia"
-    export device="$device"
-}
 
 function trim_darwin() {
     cd .repo/manifests
@@ -31,30 +24,31 @@ function trim_darwin() {
     cd ../
 }
 
-set_device
-export ROM="PixelExperience"
-manifest_url="https://github.com/PixelExperience/manifest"
 export outdir="out/target/product/$device"
-export release_repo="Akianonymus/harpia_builds"
-export ci_url="$(echo "https://cloud.drone.io/$ci_repo/"$(cat /tmp/build_no)"/1/2" | sed 's| ||')"
+ci_url="$(echo "https://cloud.drone.io/"$ci_repo"/"$(cat /tmp/build_no)"/1/2" | tr -d " ")"
 
 cd /home/ci
 
-git clone https://$GITHUB_USER:${GITHUB_TOKEN}@github.com/$GITHUB_USER/google-git-cookies.git > /dev/null 2>&1
-cd google-git-cookies
-bash setup_cookies.sh
-cd ../
-rm -rf google-git-cookies
+git clone https://"$GITHUB_USER":"${GITHUB_TOKEN}"@github.com/"$GITHUB_USER"/google-git-cookies.git > /dev/null 2>&1
+if [ -e google-git-cookies ] ; then
+    cd google-git-cookies
+    bash setup_cookies.sh
+    cd ../
+    rm -rf google-git-cookies
+else
+    echo "google-git-cookies repo not found on your account, see steps on README"
+fi
 
-mkdir $ROM
-cd $ROM
+mkdir "$ROM"
+cd "$ROM"
 
-repo init -u $manifest_url -b pie --depth 1 > /dev/null 2>&1
-echo "Sync started for $manifest_url"
-telegram -M "Sync Started for [$ROM]($manifest_url)"
+repo init -u "$manifest_url" -b "$branch" --depth 1 > /dev/null 2>&1
+echo "Sync started for "$manifest_url""
+telegram -M "Sync Started for ["$ROM"]("$manifest_url")"
 SYNC_START=$(date +"%s")
 trim_darwin >  /dev/null 2>&1
 repo sync --force-sync --current-branch --no-tags --no-clone-bundle --optimized-fetch --prune -j$(nproc --all) -q 2>&1 >>logwe 2>&1
+bash /drone/src/clone.sh
 SYNC_END=$(date +"%s")
 SYNC_DIFF=$((SYNC_END - SYNC_START))
 if [ -e frameworks/base ]; then
@@ -62,34 +56,34 @@ if [ -e frameworks/base ]; then
     echo "Build Started"
     telegram -M "Sync completed successfully in $((SYNC_DIFF / 60)) minute(s) and $((SYNC_DIFF % 60)) seconds
 
-Build Started: [See Progress]($ci_url)"
+Build Started: [See Progress]("$ci_url")"
 
     BUILD_START=$(date +"%s")
 
     . build/envsetup.sh > /dev/null 2>&1
-    set_device
-    lunch aosp_$device-userdebug > /dev/null 2>&1
-    mka bacon | grep $device
+    source /drone/src/config.sh
+    lunch "$rom_vendor_name"_"$device"-userdebug > /dev/null 2>&1
+    mka bacon | grep "$device"
     BUILD_END=$(date +"%s")
     BUILD_DIFF=$((BUILD_END - BUILD_START))
 
     export finalzip_path=$(ls "$outdir"/*201*.zip | tail -n -1)
-    export zip_name=$(echo $finalzip_path | sed "s|$outdir/||")
-    export tag=$( echo $zip_name | sed 's|.zip||')
-    if [ -e $finalzip_path ]; then
+    export zip_name=$(echo "$finalzip_path" | sed "s|"$outdir"/||")
+    export tag=$( echo "$zip_name" | sed 's|.zip||')
+    if [ -e "$finalzip_path" ]; then
         echo "Build completed successfully in $((BUILD_DIFF / 60)) minute(s) and $((BUILD_DIFF % 60)) seconds"
 
         echo "Uploading"
 
-        github-release "$release_repo" "$tag" "master" "$ROM for $device
+        github-release "$release_repo" "$tag" "master" ""$ROM" for "$device"
 
-Date: $(env TZ=Asia/Kolkata date)" "$finalzip_path"
+Date: $(env TZ="$timezone" date)" "$finalzip_path"
 
         echo "Uploaded"
 
         telegram -M "Build completed successfully in $((BUILD_DIFF / 60)) minute(s) and $((BUILD_DIFF % 60)) seconds
 
-Download: [$zip_name](https/github.com/$release_repo/releases/download/$tag/$zip_name)"
+Download: ["$zip_name"](https://github.com/"$release_repo"/releases/download/"$tag"/"$zip_name")"
 
     else
         echo "Build failed in $((BUILD_DIFF / 60)) minute(s) and $((BUILD_DIFF % 60)) seconds"
